+++ sdcm/tester.py
@@ -57,7 +57,8 @@ from .cluster_aws import ScyllaAWSCluster
 from .cluster_aws import LoaderSetAWS
 from .cluster_aws import MonitorSetAWS
 from .utils.common import get_data_dir_path, log_run_info, retrying, S3Storage, clean_cloud_instances, ScyllaCQLSession, \
-    configure_logging, get_non_system_ks_cf_list, remove_files, makedirs, format_timestamp
+    get_non_system_ks_cf_list, remove_files, makedirs, format_timestamp
+from .utils.log import configure_logging
 from . import docker
 from . import cluster_baremetal
 from . import db_stats
